requirements.txt
pytest
pyudev
psutil
Pillow
